finished flowNetworkExplorator

+++ ipa_room_exploration/common/include/ipa_room_exploration/boustrophedon_explorator.h
@@ -210,6 +210,6 @@ public:
 	// the robot should drive at.
 	void getExplorationPath(const cv::Mat& room_map, std::vector<geometry_msgs::Pose2D>& path, const float map_resolution,
 				const cv::Point starting_position, const cv::Point2d map_origin,
-				const float fitting_circle_radius, const int path_eps, const bool plan_for_footprint,
+				const float fitting_circle_radius, const double path_eps, const bool plan_for_footprint,
 				const Eigen::Matrix<float, 2, 1> robot_to_fow_vector);
 };
